Moved bearing-range error class into steam repo
